+++ src/main/kotlin/me/nobaboy/nobaaddons/NobaAddons.kt
@@ -26,7 +26,7 @@ import me.nobaboy.nobaaddons.config.NobaConfig
 import me.nobaboy.nobaaddons.config.NobaConfigUtils.safeLoad
 import me.nobaboy.nobaaddons.config.UISettings
 import me.nobaboy.nobaaddons.core.UpdateNotifier
-import me.nobaboy.nobaaddons.data.PersistentCache
+import me.nobaboy.nobaaddons.core.PersistentCache
 import me.nobaboy.nobaaddons.features.chat.alerts.IAlert
 import me.nobaboy.nobaaddons.features.chat.chatcommands.impl.DMCommands
 import me.nobaboy.nobaaddons.features.chat.chatcommands.impl.GuildCommands