+++ src/main/kotlin/me/nobaboy/nobaaddons/features/rift/RiftTimerData.kt
@@ -5,7 +5,7 @@ import dev.celestialfault.celestialconfig.Property
 import dev.celestialfault.celestialconfig.Serializer
 import me.nobaboy.nobaaddons.utils.serializers.ExtraSerializers.timestamp
 
-object RiftTimerData : ObjectProperty<RiftTimerData>("rift") {
+class RiftTimerData : ObjectProperty<RiftTimerData>("rift") {
 	var nextSplitSteal by Property.ofNullable("nextSplitSteal", serializer = Serializer.timestamp)
 	var freeRiftInfusions by Property.of("freeInfusions", 3)
 	var nextFreeInfusion by Property.ofNullable("nextFreeInfusion", serializer = Serializer.timestamp)