+++ src/main/kotlin/me/nobaboy/nobaaddons/utils/serializers/ExtraSerializers.kt
@@ -7,6 +7,7 @@ import me.nobaboy.nobaaddons.utils.NobaColor
 import me.nobaboy.nobaaddons.utils.Timestamp
 import java.awt.Color
 import java.util.EnumMap
+import java.util.UUID
 
 object ExtraSerializers {
 	inline fun <reified K : Enum<K>, V> Serializer.Companion.enumMap(valueSerializer: Serializer<V>): Serializer<EnumMap<K, V>> =
@@ -19,7 +20,7 @@ object ExtraSerializers {
 		override fun serialize(value: NobaColor): JsonElement = JsonPrimitive(value.rgb)
 
 		override fun deserialize(element: JsonElement): NobaColor? =
-			if(element is JsonPrimitive && element.isNumber) NobaColor(element.asInt) else null
+			element.let { it as? JsonPrimitive }?.takeIf { it.isNumber }?.let { NobaColor(it.asInt) }
 	}
 
 	val Serializer.Companion.timestamp get() = object : Serializer<Timestamp> {
@@ -27,5 +28,15 @@ object ExtraSerializers {
 
 		override fun deserialize(element: JsonElement): Timestamp? =
 			element.let { it as? JsonPrimitive }?.takeIf { it.isNumber }?.let { Timestamp(it.asLong) }
+	}
+
+	val Serializer.Companion.uuid get() = object : Serializer<UUID> {
+		override fun serialize(value: UUID): JsonElement = JsonPrimitive(value.toString())
+
+		override fun deserialize(element: JsonElement): UUID? =
+			element
+				.let { it as? JsonPrimitive }
+				?.takeIf { it.isString }
+				?.let { runCatching { UUID.fromString(it.asString) }.getOrNull() }
 	}
 }